Added new Time Entry container view
